deps: Fix tagged dependency matching with shrinkwraps
When you depend on a tag, eg `foo@latest`, your `npm-shrinkwrap.json` will
record this in its `_from` field.

When you install under the above circumstances the installed module will
be installed from the associated tarball URL, not by resolving the tag and
installing whatever version is now associated with it. As a result the
`_requested` section of the `package.json` will be based on the tarball URL
and cannot match the `package.json`'s tagged version.

It was _supposed_ to fallback to comparing the `package.json` `_from` field
with but it was doing this by comparing `rawSpec` to `_from`, which _can't_
work, as `rawSpec` is the requested specifier without the name (`latest` in
this case), versus `_from` (`dep@latest` in this case).

Fixes: #13496
PR-URL: https://github.com/npm/npm/pull/13941
Credit: @iarna